requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
cloudpickle==2.2.1
comm==0.2.3
contourpy==1.3.0
coverage==7.10.1
cycler==0.12.1
databricks-cli==0.18.0
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
docker==6.1.3
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.104.1
fastjsonschema==2.21.1
Flask==3.1.1
fonttools==4.59.0
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.3
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.26.0
ipython==8.18.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.7
lark==1.2.2
Mako==1.3.10
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.7.4
matplotlib-inline==0.1.7
mistune==3.1.3
mlflow==2.8.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.4
notebook_shim==0.2.4
numpy==1.24.3
oauthlib==3.3.1
overrides==7.7.0
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
pip==23.0.1
platformdirs==4.3.8
plotly==5.17.0
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==14.0.2
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==3.3.0
pytz==2023.4
PyYAML==6.0.1
pyzmq==27.0.0
qtconsole==5.6.1
QtPy==2.4.3
querystring-parser==1.2.4
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.26.0
scikit-learn==1.3.2
scipy==1.13.1
seaborn==0.13.0
Send2Trash==1.8.3
setuptools==58.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.27.0
tabulate==0.9.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.5.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250708
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.24.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xgboost==2.0.3
zipp==3.23.0
